Moved FromSetPredicate to its own file, mage.filter.predicate.mageobject.FromSetPredicate
